test: run cluster-config management API suite against the real new-model stack
ClusterConfigurationManagementApiTestBase and its two subclasses drove the
cluster-configuration-management endpoints through a hand-rolled
RecordingChangeCoordinator fake that only ever touched the legacy
ClusterConfiguration model, so it never exercised the real
ConfigurationChangeCoordinatorImpl / ClusterConfigurationManagerImpl stack
production actually runs. NewModelManagementApiEndpointsTest covered that real
stack separately, duplicating most of the same endpoint scenarios.

Move the base class and its subclasses into the io.camunda.zeebe.dynamic.config
package (needed for package-private access to ClusterConfigurationManagerImpl)
and rewire them to run against the real coordinator and manager, seeded via
CurrentClusterConfiguration.fromLegacy(...) the same way the deleted test did.
Fold NewModelManagementApiEndpointsTest's unique scenarios (cancel-change,
mode-change/routing-state physical-tenant scoping, end-to-end apply
assertions) into the base class and delete both that file and the now-unused
RecordingChangeCoordinator fake.

Running the real coordinator surfaced validation the fake had bypassed
(partition-join/leave preconditions, routing-state-required-before-scale-up,
contiguous partition IDs for purge), so a handful of fixtures needed small
realistic adjustments, each verified against the applier code enforcing it.
