+++ zeebe/dynamic-config/src/test/java/io/camunda/zeebe/dynamic/config/BareClusterConfigurationManagementApiTest.java
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.zeebe.dynamic.config.api;
+package io.camunda.zeebe.dynamic.config;
 
 import io.atomix.cluster.MemberId;
