--- zeebe/dynamic-config/src/test/java/io/camunda/zeebe/dynamic/config/NewModelManagementApiEndpointsTest.java
@@ -1,889 +1,0 @@
-/*
- * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
- * one or more contributor license agreements. See the NOTICE file distributed
- * with this work for additional information regarding copyright ownership.
- * Licensed under the Camunda License 1.0. You may not use this file
- * except in compliance with the Camunda License 1.0.
- */
-package io.camunda.zeebe.dynamic.config;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatCode;
-
-import io.atomix.cluster.MemberId;
-import io.camunda.cluster.PhysicalTenantIds;
-import io.camunda.zeebe.dynamic.config.api.ClusterConfigurationManagementRequest.AddMembersRequest;
-import io.camunda.zeebe.dynamic.config.api.ClusterConfigurationManagementRequest.BrokerScaleRequest;
-import io.camunda.zeebe.dynamic.config.api.ClusterConfigurationManagementRequest.CancelChangeRequest;
-import io.camunda.zeebe.dynamic.config.api.ClusterConfigurationManagementRequest.ClusterPatchRequest;
-import io.camunda.zeebe.dynamic.config.api.ClusterConfigurationManagementRequest.ClusterScaleRequest;
-import io.camunda.zeebe.dynamic.config.api.ClusterConfigurationManagementRequest.ExporterDeleteRequest;
-import io.camunda.zeebe.dynamic.config.api.ClusterConfigurationManagementRequest.ExporterDisableRequest;
-import io.camunda.zeebe.dynamic.config.api.ClusterConfigurationManagementRequest.ExporterEnableRequest;
-import io.camunda.zeebe.dynamic.config.api.ClusterConfigurationManagementRequest.ForceRemoveBrokersRequest;
-import io.camunda.zeebe.dynamic.config.api.ClusterConfigurationManagementRequest.JoinPartitionRequest;
-import io.camunda.zeebe.dynamic.config.api.ClusterConfigurationManagementRequest.LeavePartitionRequest;
-import io.camunda.zeebe.dynamic.config.api.ClusterConfigurationManagementRequest.ModeChangeRequest;
-import io.camunda.zeebe.dynamic.config.api.ClusterConfigurationManagementRequest.PurgeRequest;
-import io.camunda.zeebe.dynamic.config.api.ClusterConfigurationManagementRequest.ReassignPartitionsRequest;
-import io.camunda.zeebe.dynamic.config.api.ClusterConfigurationManagementRequest.RemoveMembersRequest;
-import io.camunda.zeebe.dynamic.config.api.ClusterConfigurationManagementRequest.RestoreParameters;
-import io.camunda.zeebe.dynamic.config.api.ClusterConfigurationManagementRequest.RestoreRequest;
-import io.camunda.zeebe.dynamic.config.api.ClusterConfigurationManagementRequest.RestoreResolvedRequest;
-import io.camunda.zeebe.dynamic.config.api.ClusterConfigurationManagementRequest.TenantRestoreArguments;
-import io.camunda.zeebe.dynamic.config.api.ClusterConfigurationManagementRequest.UpdatePartitionDistributorConfigRequest;
-import io.camunda.zeebe.dynamic.config.api.ClusterConfigurationManagementRequest.UpdateRoutingStateRequest;
-import io.camunda.zeebe.dynamic.config.api.ClusterConfigurationManagementRequestsHandler;
-import io.camunda.zeebe.dynamic.config.api.ClusterConfigurationRequestValidator;
-import io.camunda.zeebe.dynamic.config.changes.ClusterChangeExecutor.NoopClusterChangeExecutor;
-import io.camunda.zeebe.dynamic.config.changes.ConfigurationChangeCoordinatorImpl;
-import io.camunda.zeebe.dynamic.config.changes.GlobalConfigurationChangeAppliersImpl;
-import io.camunda.zeebe.dynamic.config.changes.ModeChangeExecutor.NoopModeChangeExecutor;
-import io.camunda.zeebe.dynamic.config.changes.NoopClusterMembershipChangeExecutor;
-import io.camunda.zeebe.dynamic.config.changes.NoopPartitionChangeExecutor;
-import io.camunda.zeebe.dynamic.config.changes.PartitionGroupConfigurationChangeAppliersImpl;
-import io.camunda.zeebe.dynamic.config.changes.PartitionScalingChangeExecutor.NoopPartitionScalingChangeExecutor;
-import io.camunda.zeebe.dynamic.config.changes.RestoreChangeExecutor.NoopRestoreChangeExecutor;
-import io.camunda.zeebe.dynamic.config.metrics.TopologyManagerMetrics;
-import io.camunda.zeebe.dynamic.config.serializer.ProtoBufSerializer;
-import io.camunda.zeebe.dynamic.config.state.BrokerPartitionState;
-import io.camunda.zeebe.dynamic.config.state.ClusterConfiguration;
-import io.camunda.zeebe.dynamic.config.state.CurrentClusterConfiguration;
-import io.camunda.zeebe.dynamic.config.state.DynamicPartitionConfig;
-import io.camunda.zeebe.dynamic.config.state.ExporterState;
-import io.camunda.zeebe.dynamic.config.state.ExporterState.State;
-import io.camunda.zeebe.dynamic.config.state.ExportingConfig;
-import io.camunda.zeebe.dynamic.config.state.ExportingState;
-import io.camunda.zeebe.dynamic.config.state.GlobalChangeOperation.MemberJoinOperation;
-import io.camunda.zeebe.dynamic.config.state.GlobalChangeOperation.MemberLeaveOperation;
-import io.camunda.zeebe.dynamic.config.state.GlobalChangeOperation.MemberRemoveOperation;
-import io.camunda.zeebe.dynamic.config.state.GlobalChangeOperation.PostScalingOperation;
-import io.camunda.zeebe.dynamic.config.state.GlobalChangeOperation.PreScalingOperation;
-import io.camunda.zeebe.dynamic.config.state.MemberState;
-import io.camunda.zeebe.dynamic.config.state.Mode;
-import io.camunda.zeebe.dynamic.config.state.PartitionDistributorConfig.RoundRobinConfig;
-import io.camunda.zeebe.dynamic.config.state.PartitionGroupConfiguration;
-import io.camunda.zeebe.dynamic.config.state.PartitionGroupOperation.PartitionChangeOperation.PartitionDeleteExporterOperation;
-import io.camunda.zeebe.dynamic.config.state.PartitionGroupOperation.PartitionChangeOperation.PartitionDisableExporterOperation;
-import io.camunda.zeebe.dynamic.config.state.PartitionGroupOperation.PartitionChangeOperation.PartitionEnableExporterOperation;
-import io.camunda.zeebe.dynamic.config.state.PartitionGroupOperation.PartitionChangeOperation.PartitionForceReconfigureOperation;
-import io.camunda.zeebe.dynamic.config.state.PartitionGroupOperation.PartitionChangeOperation.PartitionJoinOperation;
-import io.camunda.zeebe.dynamic.config.state.PartitionGroupOperation.PartitionChangeOperation.PartitionLeaveOperation;
-import io.camunda.zeebe.dynamic.config.state.PartitionGroupOperation.UpdateRoutingState;
-import io.camunda.zeebe.dynamic.config.state.PartitionState;
-import io.camunda.zeebe.dynamic.config.state.PhasedChangePlan.GlobalPhase;
-import io.camunda.zeebe.dynamic.config.state.PhasedChangePlanStatus;
-import io.camunda.zeebe.dynamic.config.state.RoutingState;
-import io.camunda.zeebe.dynamic.config.util.RequestValidatorRegistry;
-import io.camunda.zeebe.scheduler.testing.TestConcurrencyControl;
-import io.camunda.zeebe.util.Either;
-import io.micrometer.core.instrument.simple.SimpleMeterRegistry;
-import java.nio.file.Path;
-import java.time.Duration;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-import java.util.Set;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.io.TempDir;
-
-/**
- * Verifies that every {@link io.camunda.zeebe.dynamic.config.api.ClusterConfigurationManagementApi}
- * endpoint generates the expected plan for the <em>default</em> tenant when the new configuration
- * model is enabled, driven through the real request handler + coordinator + manager. Mirrors the
- * scenarios in {@code ClusterConfigurationManagementApiTest} (which runs against a recording
- * coordinator on the legacy model), asserting the same planned changes on the new path.
- */
-final class NewModelManagementApiEndpointsTest {
-
-  private static final String TENANT_B = "tenant-b";
-  private static final MemberId ID_0 = MemberId.from("0");
-  private static final MemberId ID_1 = MemberId.from("1");
-  private static final MemberId ID_2 = MemberId.from("2");
-  private static final MemberId ID_3 = MemberId.from("3");
-  private final TestConcurrencyControl executor = new TestConcurrencyControl();
-  private final DynamicPartitionConfig partitionConfig = DynamicPartitionConfig.init();
-
-  @TempDir private Path tmp;
-
-  private ClusterConfigurationManagerImpl manager;
-  private ClusterConfigurationManagementRequestsHandler handler;
-
-  /** Wires the real handler → coordinator → manager (new model) seeded from a legacy topology. */
-  private void wire(final ClusterConfiguration seed) {
-    final var persisted =
-        PersistedCurrentClusterConfiguration.ofFile(
-            tmp.resolve("config.meta"), new ProtoBufSerializer());
-    manager =
-        new ClusterConfigurationManagerImpl(
-            executor,
-            ID_0,
-            persisted,
-            new TopologyManagerMetrics(new SimpleMeterRegistry()),
-            Duration.ofMillis(1),
-            Duration.ofMillis(1));
-    manager.setCurrentConfigurationGossiper(ignored -> {});
-    manager.registerGlobalChangeAppliers(
-        new GlobalConfigurationChangeAppliersImpl(
-            new NoopClusterMembershipChangeExecutor(), new NoopClusterChangeExecutor()));
-    manager.registerPartitionGroupChangeAppliers(
-        CurrentClusterConfiguration.DEFAULT_GROUP,
-        new PartitionGroupConfigurationChangeAppliersImpl(
-            new NoopPartitionChangeExecutor(),
-            new NoopPartitionScalingChangeExecutor(),
-            new NoopModeChangeExecutor(),
-            new NoopRestoreChangeExecutor()));
-    final var coordinator = new ConfigurationChangeCoordinatorImpl(manager, ID_0, executor);
-
-    final var validatorRegistry = new RequestValidatorRegistry();
-    validatorRegistry.registerValidator(
-        null,
-        new ClusterConfigurationRequestValidator<RestoreRequest, RestoreResolvedRequest>() {
-          @Override
-          public Class<RestoreRequest> requestType() {
-            return RestoreRequest.class;
-          }
-
-          @Override
-          public Either<Exception, RestoreResolvedRequest> validate(final RestoreRequest request) {
-            return Either.right(
-                new RestoreResolvedRequest(
-                    Map.of(
-                        1,
-                        request.arguments().parameters().backupIds().stream()
-                            .mapToLong(l -> l)
-                            .toArray()),
-                    false));
-          }
-        });
-
-    handler =
-        new ClusterConfigurationManagementRequestsHandler(
-            coordinator, ID_0, executor, validatorRegistry);
-    manager
-        .updateMultiConfiguration(ignored -> CurrentClusterConfiguration.fromLegacy(seed))
-        .join();
-  }
-
-  private ClusterConfiguration singleActiveMember() {
-    return ClusterConfiguration.init().addMember(ID_0, MemberState.initializeAsActive(Map.of()));
-  }
-
-  @Test
-  void shouldPlanAddMembers() {
-    // given
-    wire(singleActiveMember());
-
-    // when
-    final var response = handler.addMembers(new AddMembersRequest(Set.of(ID_1), false)).join();
-
-    // then
-    assertThat(response.legacyResponse().plannedChanges())
-        .containsExactly(new MemberJoinOperation(ID_1));
-  }
-
-  @Test
-  void shouldPlanRemoveMembers() {
-    // given
-    wire(
-        singleActiveMember()
-            .addMember(ID_1, MemberState.initializeAsActive(Map.of()))
-            .addMember(ID_2, MemberState.initializeAsActive(Map.of())));
-
-    // when
-    final var response =
-        handler.removeMembers(new RemoveMembersRequest(Set.of(ID_1, ID_2), false)).join();
-
-    // then
-    assertThat(response.legacyResponse().plannedChanges())
-        .containsExactlyElementsOf(
-            List.of(new MemberLeaveOperation(ID_1), new MemberLeaveOperation(ID_2)));
-  }
-
-  @Test
-  void shouldPlanJoinPartition() {
-    // given — partition 1 already has an active member (ID_0), so ID_1 can join it
-    wire(
-        ClusterConfiguration.init()
-            .addMember(
-                ID_0,
-                MemberState.initializeAsActive(
-                    Map.of(1, PartitionState.active(1, partitionConfig))))
-            .addMember(ID_1, MemberState.initializeAsActive(Map.of())));
-
-    // when
-    final var response = handler.joinPartition(new JoinPartitionRequest(ID_1, 1, 3, false)).join();
-
-    // then
-    assertThat(response.legacyResponse().plannedChanges())
-        .containsExactly(new PartitionJoinOperation(ID_1, 1, 3));
-  }
-
-  @Test
-  void shouldPlanLeavePartition() {
-    // given — partition 1 has two replicas so ID_1 can leave without dropping below the minimum
-    wire(
-        ClusterConfiguration.init()
-            .addMember(
-                ID_0,
-                MemberState.initializeAsActive(
-                    Map.of(1, PartitionState.active(2, partitionConfig))))
-            .addMember(
-                ID_1,
-                MemberState.initializeAsActive(
-                    Map.of(1, PartitionState.active(1, partitionConfig)))));
-
-    // when
-    final var response = handler.leavePartition(new LeavePartitionRequest(ID_1, 1, false)).join();
-
-    // then
-    assertThat(response.legacyResponse().plannedChanges())
-        .containsExactly(new PartitionLeaveOperation(ID_1, 1, 1));
-  }
-
-  @Test
-  void shouldPlanReassignPartitions() {
-    // given
-    wire(
-        singleActiveMember()
-            .addMember(
-                ID_1,
-                MemberState.initializeAsActive(
-                    Map.of(
-                        1,
-                        PartitionState.active(1, partitionConfig),
-                        2,
-                        PartitionState.active(1, partitionConfig))))
-            .addMember(ID_2, MemberState.initializeAsActive(Map.of())));
-
-    // when
-    final var response =
-        handler.reassignPartitions(new ReassignPartitionsRequest(Set.of(ID_1, ID_2), false)).join();
-
-    // then
-    assertThat(response.legacyResponse().plannedChanges())
-        .containsExactly(
-            new PartitionJoinOperation(ID_2, 2, 1), new PartitionLeaveOperation(ID_1, 2, 1));
-  }
-
-  @Test
-  void shouldPlanScaleBrokers() {
-    // given
-    wire(
-        singleActiveMember()
-            .updateMember(ID_0, m -> m.addPartition(1, PartitionState.active(1, partitionConfig)))
-            .updateMember(ID_0, m -> m.addPartition(2, PartitionState.active(1, partitionConfig))));
-
-    // when
-    final var response =
-        handler.scaleMembers(new BrokerScaleRequest(Set.of(ID_0, ID_1), false)).join();
-
-    // then
-    assertThat(response.legacyResponse().plannedChanges())
-        .containsExactly(
-            new PreScalingOperation(ID_0, Set.of(ID_0, ID_1)),
-            new MemberJoinOperation(ID_1),
-            new PartitionJoinOperation(ID_1, 2, 1),
-            new PartitionLeaveOperation(ID_0, 2, 1),
-            new PostScalingOperation(ID_0, Set.of(ID_0, ID_1)));
-  }
-
-  @Test
-  void shouldPlanScaleCluster() {
-    // given
-    wire(
-        singleActiveMember()
-            .updateMember(ID_0, m -> m.addPartition(1, PartitionState.active(1, partitionConfig)))
-            .updateMember(ID_0, m -> m.addPartition(2, PartitionState.active(1, partitionConfig))));
-
-    // when — scale the cluster to two brokers, keeping the current partition count (no scale-up)
-    final var response =
-        handler
-            .scaleCluster(
-                new ClusterScaleRequest(
-                    Optional.of(2), Optional.of(2), Optional.empty(), Optional.empty(), false))
-            .join();
-
-    // then
-    assertThat(response.legacyResponse().plannedChanges())
-        .startsWith(new PreScalingOperation(ID_0, Set.of(ID_0, ID_1)))
-        .endsWith(new PostScalingOperation(ID_0, Set.of(ID_0, ID_1)))
-        .contains(new MemberJoinOperation(ID_1));
-  }
-
-  @Test
-  void shouldPlanPatchCluster() {
-    // given
-    wire(
-        singleActiveMember()
-            .updateMember(ID_0, m -> m.addPartition(1, PartitionState.active(1, partitionConfig)))
-            .updateMember(ID_0, m -> m.addPartition(2, PartitionState.active(1, partitionConfig))));
-
-    // when — add a broker, keeping the current partition count (no scale-up)
-    final var response =
-        handler
-            .patchCluster(
-                new ClusterPatchRequest(
-                    Set.of(ID_1), Set.of(), Optional.of(2), Optional.empty(), false))
-            .join();
-
-    // then
-    assertThat(response.legacyResponse().plannedChanges())
-        .startsWith(new PreScalingOperation(ID_0, Set.of(ID_0, ID_1)))
-        .endsWith(new PostScalingOperation(ID_0, Set.of(ID_0, ID_1)))
-        .contains(new MemberJoinOperation(ID_1));
-  }
-
-  @Test
-  void shouldApplyScaleClusterToOnlyTheRequestedPhysicalTenant() {
-    // given — both physical tenants start with 1 partition
-    wireTwoPhysicalTenants();
-    seedRoutingState(1, CurrentClusterConfiguration.DEFAULT_GROUP, TENANT_B);
-
-    // when — scale only tenant-b's partition count up to 2
-    handler
-        .scaleCluster(
-            new ClusterScaleRequest(
-                Optional.empty(),
-                Optional.of(2),
-                Optional.empty(),
-                Optional.empty(),
-                Optional.of(TENANT_B),
-                false))
-        .join();
-
-    // then — only tenant-b's partition count changed; the default group is untouched
-    final var config = manager.getMultiConfiguration().join();
-    assertThat(config.phasedChangeState().pending()).isEmpty();
-    assertThat(config.partitionGroup(TENANT_B).partitionCount()).isEqualTo(2);
-    assertThat(config.partitionGroup(CurrentClusterConfiguration.DEFAULT_GROUP).partitionCount())
-        .isEqualTo(1);
-  }
-
-  @Test
-  void shouldApplyScaleClusterToOnlyTheDefaultTenantWhenUnscoped() {
-    // given — both physical tenants start with 1 partition
-    wireTwoPhysicalTenants();
-    seedRoutingState(1, CurrentClusterConfiguration.DEFAULT_GROUP, TENANT_B);
-
-    // when — no physicalTenant parameter is given
-    handler
-        .scaleCluster(
-            new ClusterScaleRequest(
-                Optional.empty(), Optional.of(2), Optional.empty(), Optional.empty(), false))
-        .join();
-
-    // then — only the default tenant's partition count changed
-    final var config = manager.getMultiConfiguration().join();
-    assertThat(config.phasedChangeState().pending()).isEmpty();
-    assertThat(config.partitionGroup(CurrentClusterConfiguration.DEFAULT_GROUP).partitionCount())
-        .isEqualTo(2);
-    assertThat(config.partitionGroup(TENANT_B).partitionCount()).isEqualTo(1);
-  }
-
-  @Test
-  void shouldApplyPatchClusterToOnlyTheRequestedPhysicalTenant() {
-    // given — both physical tenants start with 1 partition
-    wireTwoPhysicalTenants();
-    seedRoutingState(1, CurrentClusterConfiguration.DEFAULT_GROUP, TENANT_B);
-
-    // when — patch only tenant-b's partition count up to 2
-    handler
-        .patchCluster(
-            new ClusterPatchRequest(
-                Set.of(), Set.of(), Optional.of(2), Optional.empty(), Optional.of(TENANT_B), false))
-        .join();
-
-    // then — only tenant-b's partition count changed; the default group is untouched
-    final var config = manager.getMultiConfiguration().join();
-    assertThat(config.phasedChangeState().pending()).isEmpty();
-    assertThat(config.partitionGroup(TENANT_B).partitionCount()).isEqualTo(2);
-    assertThat(config.partitionGroup(CurrentClusterConfiguration.DEFAULT_GROUP).partitionCount())
-        .isEqualTo(1);
-  }
-
-  @Test
-  void shouldApplyPatchClusterToOnlyTheDefaultTenantWhenUnscoped() {
-    // given — both physical tenants start with 1 partition
-    wireTwoPhysicalTenants();
-    seedRoutingState(1, CurrentClusterConfiguration.DEFAULT_GROUP, TENANT_B);
-
-    // when — no physicalTenant parameter is given
-    handler
-        .patchCluster(
-            new ClusterPatchRequest(Set.of(), Set.of(), Optional.of(2), Optional.empty(), false))
-        .join();
-
-    // then — only the default tenant's partition count changed
-    final var config = manager.getMultiConfiguration().join();
-    assertThat(config.phasedChangeState().pending()).isEmpty();
-    assertThat(config.partitionGroup(CurrentClusterConfiguration.DEFAULT_GROUP).partitionCount())
-        .isEqualTo(2);
-    assertThat(config.partitionGroup(TENANT_B).partitionCount()).isEqualTo(1);
-  }
-
-  @Test
-  void shouldPlanForceScaleDown() {
-    // given
-    wire(fourMemberTwoPartitionCluster());
-
-    // when
-    final var response =
-        handler.forceScaleDown(new BrokerScaleRequest(Set.of(ID_0, ID_2), false)).join();
-
-    // then
-    assertThat(response.legacyResponse().plannedChanges())
-        .containsExactlyInAnyOrder(
-            new PartitionForceReconfigureOperation(ID_0, 1, Set.of(ID_0)),
-            new PartitionForceReconfigureOperation(ID_2, 2, Set.of(ID_2)),
-            new MemberRemoveOperation(ID_0, ID_1),
-            new MemberRemoveOperation(ID_0, ID_3));
-  }
-
-  @Test
-  void shouldPlanForceRemoveBrokers() {
-    // given
-    wire(fourMemberTwoPartitionCluster());
-
-    // when
-    final var response =
-        handler.forceRemoveBrokers(new ForceRemoveBrokersRequest(Set.of(ID_1, ID_3), false)).join();
-
-    // then
-    assertThat(response.legacyResponse().plannedChanges())
-        .containsExactlyInAnyOrder(
-            new PartitionForceReconfigureOperation(ID_0, 1, Set.of(ID_0)),
-            new PartitionForceReconfigureOperation(ID_2, 2, Set.of(ID_2)),
-            new MemberRemoveOperation(ID_0, ID_1),
-            new MemberRemoveOperation(ID_0, ID_3));
-  }
-
-  @Test
-  void shouldPlanDisableExporter() {
-    // given
-    final var exporterId = "exporterId";
-    final var partitionConfigWithExporter =
-        new DynamicPartitionConfig(
-            new ExportingConfig(
-                ExportingState.EXPORTING,
-                Map.of(exporterId, new ExporterState(1, State.ENABLED, Optional.empty()))));
-    wire(
-        singleActiveMember()
-            .updateMember(
-                ID_0,
-                m -> m.addPartition(1, PartitionState.active(1, partitionConfigWithExporter))));
-
-    // when
-    final var response =
-        handler
-            .disableExporter(new ExporterDisableRequest(exporterId, Optional.empty(), false))
-            .join();
-
-    // then
-    assertThat(response.legacyResponse().plannedChanges())
-        .containsExactly(new PartitionDisableExporterOperation(ID_0, 1, exporterId));
-  }
-
-  @Test
-  void shouldApplyDisableExporterEndToEndForLocalMember() {
-    // given — the exporter lives on the local member's partition
-    final var exporterId = "exporterId";
-    final var partitionConfigWithExporter =
-        new DynamicPartitionConfig(
-            new ExportingConfig(
-                ExportingState.EXPORTING,
-                Map.of(exporterId, new ExporterState(1, State.ENABLED, Optional.empty()))));
-    wire(
-        singleActiveMember()
-            .updateMember(
-                ID_0,
-                m -> m.addPartition(1, PartitionState.active(1, partitionConfigWithExporter))));
-
-    // when
-    handler.disableExporter(new ExporterDisableRequest(exporterId, Optional.empty(), false)).join();
-
-    // then — the plan is applied on the local member and completes
-    final var config = manager.getMultiConfiguration().join();
-    assertThat(config.phasedChangeState().pending()).isEmpty();
-    final var exporterState =
-        config
-            .partitionGroup(CurrentClusterConfiguration.DEFAULT_GROUP)
-            .getMember(ID_0)
-            .partitions()
-            .get(1)
-            .config()
-            .exporting()
-            .exporters()
-            .get(exporterId);
-    assertThat(exporterState.state()).isEqualTo(State.DISABLED);
-  }
-
-  @Test
-  void shouldPlanDeleteExporter() {
-    // given — the exporter is in CONFIG_NOT_FOUND, the state a delete requires
-    final var exporterId = "exporterId";
-    final var configWithExporter =
-        new DynamicPartitionConfig(
-            new ExportingConfig(
-                ExportingState.EXPORTING,
-                Map.of(
-                    exporterId, new ExporterState(1, State.CONFIG_NOT_FOUND, Optional.empty()))));
-    wire(
-        singleActiveMember()
-            .updateMember(
-                ID_0, m -> m.addPartition(1, PartitionState.active(1, configWithExporter))));
-
-    // when
-    final var response =
-        handler
-            .deleteExporter(new ExporterDeleteRequest(exporterId, Optional.empty(), false))
-            .join();
-
-    // then
-    assertThat(response.legacyResponse().plannedChanges())
-        .containsExactly(new PartitionDeleteExporterOperation(ID_0, 1, exporterId));
-  }
-
-  @Test
-  void shouldPlanEnableExporter() {
-    // given — the exporter is DISABLED so it can be enabled
-    final var exporterId = "exporterId";
-    final var configWithExporter =
-        new DynamicPartitionConfig(
-            new ExportingConfig(
-                ExportingState.EXPORTING,
-                Map.of(exporterId, new ExporterState(1, State.DISABLED, Optional.empty()))));
-    wire(
-        singleActiveMember()
-            .updateMember(
-                ID_0, m -> m.addPartition(1, PartitionState.active(1, configWithExporter))));
-
-    // when
-    final var response =
-        handler
-            .enableExporter(
-                new ExporterEnableRequest(exporterId, Optional.empty(), Optional.empty(), false))
-            .join();
-
-    // then
-    assertThat(response.legacyResponse().plannedChanges())
-        .containsExactly(
-            new PartitionEnableExporterOperation(ID_0, 1, exporterId, Optional.empty()));
-  }
-
-  @Test
-  void shouldPlanPurge() {
-    // given — a member with a partition
-    wire(
-        singleActiveMember()
-            .updateMember(ID_0, m -> m.addPartition(1, PartitionState.active(1, partitionConfig))));
-
-    // when
-    final var response = handler.purge(new PurgeRequest(Optional.empty(), false)).join();
-
-    // then — purge produces a plan and preserves the partition assignment of every member
-    assertThat(response.legacyResponse().plannedChanges()).isNotEmpty();
-    assertThat(response.legacyResponse().expectedConfiguration())
-        .containsOnlyKeys(
-            response.legacyResponse().currentConfiguration().keySet().toArray(MemberId[]::new));
-  }
-
-  @Test
-  void shouldApplyModeChangeEndToEndForLocalMember() {
-    // given — the local member is active and processing in the default group
-    wire(
-        ClusterConfiguration.init()
-            .addMember(
-                ID_0,
-                MemberState.initializeAsActive(
-                    Map.of(1, PartitionState.active(1, partitionConfig)))));
-
-    // when — switch to recovery mode
-    handler
-        .modeChange(
-            new ModeChangeRequest(
-                Optional.of(CurrentClusterConfiguration.DEFAULT_GROUP), Mode.RECOVERING, false))
-        .join();
-
-    // then — the plan is applied and the local member is now in recovery mode
-    final var config = manager.getMultiConfiguration().join();
-    assertThat(config.phasedChangeState().pending()).isEmpty();
-    assertThat(
-            config.partitionGroup(CurrentClusterConfiguration.DEFAULT_GROUP).getMember(ID_0).mode())
-        .isEqualTo(Mode.RECOVERING);
-  }
-
-  @Test
-  void shouldApplyModeChangeToEveryPhysicalTenantInOnePlan() {
-    // given
-    wireTwoPhysicalTenants();
-
-    // when
-    handler.modeChange(new ModeChangeRequest(Optional.empty(), Mode.RECOVERING, false)).join();
-
-    // then
-    final var config = manager.getMultiConfiguration().join();
-    assertThat(config.phasedChangeState().pending()).isEmpty();
-    assertThat(
-            config.partitionGroup(CurrentClusterConfiguration.DEFAULT_GROUP).getMember(ID_0).mode())
-        .isEqualTo(Mode.RECOVERING);
-    assertThat(config.partitionGroup(TENANT_B).getMember(ID_0).mode()).isEqualTo(Mode.RECOVERING);
-  }
-
-  @Test
-  void shouldApplyModeChangeToOnlyTheRequestedPhysicalTenant() {
-    // given
-    wireTwoPhysicalTenants();
-
-    // when
-    handler.modeChange(new ModeChangeRequest(Optional.of(TENANT_B), Mode.RECOVERING, false)).join();
-
-    // then
-    final var config = manager.getMultiConfiguration().join();
-    assertThat(config.phasedChangeState().pending()).isEmpty();
-    assertThat(config.partitionGroup(TENANT_B).getMember(ID_0).mode()).isEqualTo(Mode.RECOVERING);
-    assertThat(
-            config.partitionGroup(CurrentClusterConfiguration.DEFAULT_GROUP).getMember(ID_0).mode())
-        .isEqualTo(Mode.PROCESSING);
-  }
-
-  /**
-   * Wires a cluster with two physical tenants — the default group and {@link #TENANT_B} — both
-   * hosted by {@link #ID_0} and both processing.
-   */
-  private void wireTwoPhysicalTenants() {
-    wire(
-        ClusterConfiguration.init()
-            .addMember(
-                ID_0,
-                MemberState.initializeAsActive(
-                    Map.of(1, PartitionState.active(1, partitionConfig)))));
-    manager.registerPartitionGroupChangeAppliers(
-        TENANT_B,
-        new PartitionGroupConfigurationChangeAppliersImpl(
-            new NoopPartitionChangeExecutor(),
-            new NoopPartitionScalingChangeExecutor(),
-            new NoopModeChangeExecutor(),
-            new NoopRestoreChangeExecutor()));
-    manager
-        .updateMultiConfiguration(
-            current ->
-                new CurrentClusterConfiguration(
-                    current.version(),
-                    current.globalConfiguration(),
-                    Map.of(
-                        CurrentClusterConfiguration.DEFAULT_GROUP,
-                        current.partitionGroup(CurrentClusterConfiguration.DEFAULT_GROUP),
-                        TENANT_B,
-                        new PartitionGroupConfiguration(
-                            1,
-                            0,
-                            Map.of(
-                                ID_0,
-                                BrokerPartitionState.initialize(
-                                    Map.of(1, PartitionState.active(1, partitionConfig)))),
-                            Optional.empty(),
-                            Optional.empty(),
-                            Optional.empty())),
-                    current.phasedChangeState()))
-        .join();
-  }
-
-  /**
-   * Seeds a routing state on each named group: {@code StartPartitionScaleUpApplier} refuses to
-   * scale up a group whose routing state is not initialized yet, so the scale-up tests need it
-   * seeded on both groups regardless of which one is actually targeted.
-   */
-  private void seedRoutingState(final int partitionCount, final String... groupIds) {
-    manager
-        .updateMultiConfiguration(
-            current -> {
-              var updated = current;
-              for (final var groupId : groupIds) {
-                updated =
-                    updated.updatePartitionGroupConfig(
-                        groupId,
-                        group ->
-                            group.setRoutingState(
-                                RoutingState.initializeWithPartitionCount(partitionCount)));
-              }
-              return updated;
-            })
-        .join();
-  }
-
-  @Test
-  void shouldPlanUpdateRoutingState() {
-    // given
-    wire(
-        singleActiveMember()
-            .updateMember(ID_0, m -> m.addPartition(1, PartitionState.active(1, partitionConfig))));
-    final var routingState = RoutingState.initializeWithPartitionCount(1);
-
-    // when
-    final var response =
-        handler
-            .updateRoutingState(
-                new UpdateRoutingStateRequest(Optional.of(routingState), Optional.empty(), false))
-            .join();
-
-    // then
-    assertThat(response.legacyResponse().plannedChanges())
-        .containsExactly(new UpdateRoutingState(ID_0, Optional.of(routingState)));
-  }
-
-  @Test
-  void shouldApplyRoutingStateUpdateToOnlyTheRequestedPhysicalTenant() {
-    // given
-    wireTwoPhysicalTenants();
-    final var routingState = RoutingState.initializeWithPartitionCount(1);
-
-    // when
-    handler
-        .updateRoutingState(
-            new UpdateRoutingStateRequest(Optional.of(routingState), Optional.of(TENANT_B), false))
-        .join();
-
-    // then — only tenant-b's routing state changed; the default group's is left untouched
-    final var config = manager.getMultiConfiguration().join();
-    assertThat(config.phasedChangeState().pending()).isEmpty();
-    assertThat(config.partitionGroup(TENANT_B).routingState()).contains(routingState);
-    assertThat(config.partitionGroup(CurrentClusterConfiguration.DEFAULT_GROUP).routingState())
-        .isEmpty();
-  }
-
-  @Test
-  void shouldApplyRoutingStateUpdateToTheDefaultPhysicalTenantWhenUnscoped() {
-    // given — unlike mode changes, where an absent physicalTenantId means "every tenant", an
-    // unscoped routing-state update keeps writing only the default group
-    wireTwoPhysicalTenants();
-    final var routingState = RoutingState.initializeWithPartitionCount(1);
-
-    // when
-    handler
-        .updateRoutingState(
-            new UpdateRoutingStateRequest(Optional.of(routingState), Optional.empty(), false))
-        .join();
-
-    // then
-    final var config = manager.getMultiConfiguration().join();
-    assertThat(config.phasedChangeState().pending()).isEmpty();
-    assertThat(config.partitionGroup(CurrentClusterConfiguration.DEFAULT_GROUP).routingState())
-        .contains(routingState);
-    assertThat(config.partitionGroup(TENANT_B).routingState()).isEmpty();
-  }
-
-  @Test
-  void shouldRejectUpdatePartitionDistributionForNonZoneAwareConfig() {
-    // given — a bare cluster and a non-zone-aware config, which the transformer rejects
-    wire(
-        singleActiveMember()
-            .updateMember(ID_0, m -> m.addPartition(1, PartitionState.active(1, partitionConfig))));
-
-    // when / then — the endpoint is reachable on the new model and rejects the invalid request
-    assertThatCode(
-            () ->
-                handler
-                    .updatePartitionDistribution(
-                        new UpdatePartitionDistributorConfigRequest(new RoundRobinConfig(), false))
-                    .join())
-        .hasMessageContaining("ZONE_AWARE");
-  }
-
-  @Test
-  void shouldValidateRestoreOnlyWhenRecovering() {
-    // given — the local member is in recovery mode. It must host a partition: recovery mode lives
-    // on BrokerPartitionState, so a partitionless member is not in the default group and its mode
-    // would not survive the round-trip through toLegacyDefault.
-    wire(
-        ClusterConfiguration.init()
-            .addMember(
-                ID_0,
-                MemberState.initializeAsActive(Map.of(1, PartitionState.active(1, partitionConfig)))
-                    .toRecovering()));
-
-    // when — a dry-run restore is validated
-    final var result =
-        handler.restore(
-            new RestoreRequest(
-                PhysicalTenantIds.DEFAULT_PHYSICAL_TENANT_ID,
-                new TenantRestoreArguments(
-                    new RestoreParameters(List.of(100L), null, null), "elasticsearch", false),
-                true));
-
-    // then — accepted, because toLegacyDefault projects the member back to RECOVERING
-    assertThatCode(result::join).doesNotThrowAnyException();
-  }
-
-  @Test
-  void shouldRejectRestoreWhenNotRecovering() {
-    // given — the local member is active (not recovering)
-    wire(singleActiveMember());
-
-    // when / then — restore is rejected
-    assertThatCode(
-            () ->
-                handler
-                    .restore(
-                        new RestoreRequest(
-                            PhysicalTenantIds.DEFAULT_PHYSICAL_TENANT_ID,
-                            new TenantRestoreArguments(
-                                new RestoreParameters(List.of(100L), null, null),
-                                "elasticsearch",
-                                false),
-                            false))
-                    .join())
-        .hasMessageContaining("recovery mode");
-  }
-
-  @Test
-  void shouldCancelTopologyChangeEndToEnd() {
-    // given — a pending plan whose operation targets a member other than the local one, so it
-    // never drains on its own
-    wire(
-        singleActiveMember()
-            .addMember(ID_1, MemberState.initializeAsActive(Map.of()))
-            .updateMember(ID_0, m -> m.addPartition(1, PartitionState.active(1, partitionConfig))));
-    manager
-        .updateMultiConfiguration(
-            c -> c.initPlan(List.of(new GlobalPhase(List.of(new MemberLeaveOperation(ID_1))))))
-        .join();
-    final var changeId =
-        manager.getMultiConfiguration().join().phasedChangeState().onlyPending().id();
-
-    // when — cancelled via the management API endpoint
-    final var response = handler.cancelTopologyChange(new CancelChangeRequest(changeId)).join();
-
-    // then — no pending change remains, on both the response and the real configuration
-    assertThat(response.phasedChangeState().pending()).isEmpty();
-    final var config = manager.getMultiConfiguration().join();
-    assertThat(config.phasedChangeState().pending()).isEmpty();
-    assertThat(config.phasedChangeState().lastChange())
-        .hasValueSatisfying(
-            last -> assertThat(last.status()).isEqualTo(PhasedChangePlanStatus.CANCELLED));
-  }
-
-  @Test
-  void shouldGetTopology() {
-    // given
-    wire(singleActiveMember().addMember(ID_1, MemberState.initializeAsActive(Map.of())));
-
-    // when
-    final var topology = handler.getTopology().join();
-
-    // then — the global configuration carries all cluster members
-    assertThat(topology.globalConfiguration().members()).containsKeys(ID_0, ID_1);
-  }
-
-  private ClusterConfiguration fourMemberTwoPartitionCluster() {
-    return ClusterConfiguration.init()
-        .addMember(ID_0, MemberState.initializeAsActive(Map.of()))
-        .addMember(ID_1, MemberState.initializeAsActive(Map.of()))
-        .addMember(ID_2, MemberState.initializeAsActive(Map.of()))
-        .addMember(ID_3, MemberState.initializeAsActive(Map.of()))
-        .updateMember(ID_0, m -> m.addPartition(1, PartitionState.active(1, partitionConfig)))
-        .updateMember(ID_1, m -> m.addPartition(1, PartitionState.active(2, partitionConfig)))
-        .updateMember(ID_2, m -> m.addPartition(2, PartitionState.active(1, partitionConfig)))
-        .updateMember(ID_3, m -> m.addPartition(2, PartitionState.active(2, partitionConfig)));
-  }
-}
